refactored favoriting button to its own component
